Split up geometries and features into separate files
