+++ Sources/Turf/GeoJSON.swift
@@ -42,8 +42,6 @@ extension FeatureIdentifier: Codable {
     }
 }
 
-public protocol GeometryObject: Codable { }
-
 public protocol GeoJSONObject: Codable {
     var identifier: FeatureIdentifier? { get set }
     var properties: [String: AnyJSONType]? { get set }
@@ -82,69 +80,12 @@ public struct Feature: Codable {
     }
 }
 
-public struct Geometry: Codable {
-    public var type: String
-    
-    public var geometryType: GeometryType? {
-        return GeometryType(rawValue: type)
-    }
-}
-
-// Polyline has been renamed to `LineString`. This alias is for backwards compatibility.
-public typealias Polyline = LineString
-
-/**
- A `LineString` struct represents a shape consisting of two or more coordinates,
- specified as `[CLLocationCoordinate2D]`
- */
-public struct LineString: Codable {
-    var type: String = GeometryType.LineString.rawValue
-    var coordinates: [CLLocationCoordinate2D]
-}
-
-public struct Polygon: Codable {
-    var type: String = GeometryType.Polygon.rawValue
-    var coordinates: [[CLLocationCoordinate2D]]
-    
-    init(coordinates: [[CLLocationCoordinate2D]]) {
-        self.coordinates = coordinates
-    }
-    
-    var innerRings: [Ring]? {
-        get { return Array(coordinates.suffix(from: 1)).map { Ring(coordinates: $0) } }
-    }
-    
-    var outerRing: Ring {
-        get { return Ring(coordinates: coordinates.first! ) }
-    }
-}
-
-public struct Point: Codable {
-    var type: String = GeometryType.Point.rawValue
-    var coordinates: CLLocationCoordinate2D
-}
-
-public struct MultiPoint: Codable {
-    var type: String = GeometryType.MultiPoint.rawValue
-    var coordinates: [CLLocationCoordinate2D]
-}
-
-public struct MultiLineString: Codable {
-    var type: String = GeometryType.MultiLineString.rawValue
-    var coordinates: [[CLLocationCoordinate2D]]
-}
-
-public struct MultiPolygon: Codable {
-    var type: String = GeometryType.MultiLineString.rawValue
-    var coordinates: [[[CLLocationCoordinate2D]]]
-}
-
-public class PointFeature: GeoJSONObject {
+public struct PointFeature: GeoJSONObject {
     public var identifier: FeatureIdentifier?
     public var geometry: Point!
     public var properties: [String : AnyJSONType]?
     
-    public required init(from decoder: Decoder) throws {
+    public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: GeoJSONCodingKeys.self)
         geometry = try container.decode(Point.self, forKey: .geometry)
         properties = try container.decode([String: AnyJSONType]?.self, forKey: .properties)
@@ -159,7 +100,7 @@ public class PointFeature: GeoJSONObject {
     }
 }
 
-public struct LineStringFeature: GeoJSONObject, GeometryObject {
+public struct LineStringFeature: GeoJSONObject {
     public var identifier: FeatureIdentifier?
     public var geometry: LineString!
     public var properties: [String : AnyJSONType]?
@@ -282,8 +223,8 @@ public struct FeatureCollection: GeoJSONObject {
             while (!unkeyedContainer.isAtEnd) {
                 let feature = try unkeyedContainer.decode(Feature.self)
 
-                if let test = feature.simplifiedGeometry?.geometryType {
-                    switch test {
+                if let geometryType = feature.simplifiedGeometry?.geometryType {
+                    switch geometryType {
                     case .LineString:
                         featureTypes.append(LineStringFeature.self)
                     case .MultiLineString:
@@ -330,17 +271,6 @@ public struct FeatureCollection: GeoJSONObject {
         //try container.encode(features, forKey: .features)
         try container.encode(properties, forKey: .properties)
     }
-}
-
-public enum GeometryType: String {
-    case Point
-    case LineString
-    case Polygon
-    case MultiPoint
-    case MultiLineString
-    case MultiPolygon
-    
-    static let allValues: [GeometryType] = [.Point, .LineString, .Polygon, .MultiPoint, .MultiLineString, .MultiPolygon]
 }
 
 public enum GeoJSONType: String {
